requirements.txt
pandas
openpyxl
matplotlib
seaborn
jinja2
streamlit
PyYAML
scipy
